backend/requirements.txt
# Pronto Casa Backend

fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
celery[redis]==5.3.6
redis==5.0.1
boto3==1.34.25
stripe==7.10.0
twilio==8.13.0
firebase-admin==6.4.0
Pillow==10.2.0
openai==1.10.0
geopy==2.4.1
geoalchemy2==0.14.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
